Update Package and Rearrange Exception

+++ PidgeotMailMVVM/Lib/GoogleService.cs
@@ -22,7 +22,7 @@ namespace PidgeotMail.Lib
 		private static readonly string TextCredential = "{\"installed\":{\"client_id\":\"155023122563-9qnq9022jb9189377rjt98k91s393pen.apps.googleusercontent.com\",\"project_id\":\"PidgeotMail\",\"auth_uri\":\"https://accounts.google.com/o/oauth2/auth\",\"token_uri\":\"https://oauth2.googleapis.com/token\",\"auth_provider_x509_cert_url\":\"https://www.googleapis.com/oauth2/v1/certs\",\"client_secret\":\"ztuBZE_qYRBEJbYSAgbPimu4\",\"redirect_uris\":[\"urn:ietf:wg:oauth:2.0:oob\",\"http://localhost\"]}}";
 		public static UserCredential Credential => credential;
 
-		private static string[] Scopes = {GmailService.Scope.GmailReadonly, SheetsService.Scope.Spreadsheets, "https://mail.google.com/" };
+		private static string[] Scopes = { GmailService.Scope.GmailReadonly, SheetsService.Scope.Spreadsheets, "https://mail.google.com/" };
 		private static readonly log4net.ILog log = log4net.LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
 
 		private static Stream GenerateStreamFromString(string s)
@@ -69,9 +69,10 @@ namespace PidgeotMail.Lib
 						File.Delete("credentials.json");
 					}
 				}
-				catch
+				catch (Exception e)
 				{
-					Directory.Delete(UserSettings.TokenFolder, true);
+					if (Directory.Exists(UserSettings.TokenFolder))Directory.Delete(UserSettings.TokenFolder, true);
+					throw e;
 				}
 			}
 			);
